+++ src/folder/data-access/folder/useGetFolder.ts
@@ -1,9 +1,9 @@
-import { SampleFolderRawData } from "@/src/folder/type";
-import { mapFolderData } from "@/src/folder/util/map";
+import { SampleFolderRawData, mapFolderData } from "@/src/folder";
 import { useAsync, axiosInstance } from "@/src/util";
 
 export const useGetFolder = () => {
-  const getFolder = () => axiosInstance.get<{ folder: SampleFolderRawData }>("sample/folder");
+  const getFolder = () =>
+    axiosInstance.get<{ folder: SampleFolderRawData }>("sample/folder");
   const { loading, error, data } = useAsync(getFolder);
 
   const folderData = mapFolderData(data?.folder);
